Added new Address resource.
