+++ src/main/java/develop/gym/dto/GroupTrainingDto.java
@@ -1,4 +1,4 @@
-package develop.dto;
+package develop.gym.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
