+++ src/main/java/develop/gym/exception/CustomerNotFoundException.java
@@ -1,4 +1,4 @@
-package develop.exception;
+package develop.gym.exception;
 
 public class CustomerNotFoundException extends Exception {
     public CustomerNotFoundException(String msg){super(msg);}
